Matter Lock: Switch sub-driver and main-driver to support SLGA migration (#2842)
---------
