+++ drivers/SmartThings/matter-lock/src/test/test_matter_lock_battery.lua
@@ -67,9 +67,7 @@ local function test_init()
   test.socket.device_lifecycle:__queue_receive({ mock_device.id, "init" })
   local subscribe_request = clusters.DoorLock.attributes.LockState:subscribe(mock_device)
   subscribe_request:merge(clusters.PowerSource.attributes.BatPercentRemaining:subscribe(mock_device))
-  subscribe_request:merge(clusters.DoorLock.events.DoorLockAlarm:subscribe(mock_device))
   subscribe_request:merge(clusters.DoorLock.events.LockOperation:subscribe(mock_device))
-  subscribe_request:merge(clusters.DoorLock.events.LockUserChange:subscribe(mock_device))
   test.socket["matter"]:__expect_send({mock_device.id, subscribe_request})
   test.socket.device_lifecycle:__queue_receive({ mock_device.id, "doConfigure" })
   mock_device:expect_metadata_update({ provisioning_state = "PROVISIONED" })
@@ -88,9 +86,7 @@ local function test_init_no_battery()
   test.socket.device_lifecycle:__queue_receive({ mock_device_no_battery.id, "init" })
   local subscribe_request = clusters.DoorLock.attributes.LockState:subscribe(mock_device_no_battery)
   subscribe_request:merge(clusters.PowerSource.attributes.BatPercentRemaining:subscribe(mock_device_no_battery))
-  subscribe_request:merge(clusters.DoorLock.events.DoorLockAlarm:subscribe(mock_device_no_battery))
   subscribe_request:merge(clusters.DoorLock.events.LockOperation:subscribe(mock_device_no_battery))
-  subscribe_request:merge(clusters.DoorLock.events.LockUserChange:subscribe(mock_device_no_battery))
   test.socket["matter"]:__expect_send({mock_device_no_battery.id, subscribe_request})
   test.socket.device_lifecycle:__queue_receive({ mock_device_no_battery.id, "doConfigure" })
   mock_device_no_battery:expect_metadata_update({ profile = "base-lock-nobattery" })
