+++ drivers/SmartThings/matter-lock/src/test/test_matter_lock_modular.lua
@@ -625,7 +625,7 @@ test.register_coroutine_test(
 
 
 local mock_nuki_smart_lock_ultra = test.mock_device.build_test_matter_device({
-  profile = t_utils.get_profile_definition("lock-nocodes-notamper.yml"),
+  profile = t_utils.get_profile_definition("lock-unlatch.yml"),
   manufacturer_info = {
     vendor_id = 0x135D,
     product_id = 0x00A1,
