+++ drivers/SmartThings/matter-lock/src/test/test_new_matter_lock.lua
@@ -1,7 +1,7 @@
 -- Copyright 2023 SmartThings, Inc.
 -- Licensed under the Apache License, Version 2.0
 
-
+-- These are test units designed to validate User, Credential, and Schedule features.
 local test = require "integration_test"
 test.set_rpc_version(0)
 local capabilities = require "st.capabilities"
@@ -13,7 +13,7 @@ local types = DoorLock.types
 local lock_utils = require "lock_utils"
 
 local mock_device = test.mock_device.build_test_matter_device({
-  profile = t_utils.get_profile_definition("lock-user-pin-schedule.yml"),
+  profile = t_utils.get_profile_definition("lock.yml"),
   manufacturer_info = {
     vendor_id = 0x115f,
     product_id = 0x2802,
@@ -48,6 +48,12 @@ local mock_device = test.mock_device.build_test_matter_device({
 local DoorLockFeatureMapAttr = {ID = 0xFFFC, cluster = DoorLock.ID}
 local function test_init()
   test.disable_startup_messages()
+  -- device_added
+  test.mock_device.add_test_device(mock_device)
+  test.socket.device_lifecycle:__queue_receive({ mock_device.id, "added" })
+  test.socket.capability:__expect_send(
+    mock_device:generate_test_message("main", capabilities.lockAlarm.alarm.clear({state_change = true}))
+  )
   -- subscribe request
   local subscribe_request = DoorLock.attributes.LockState:subscribe(mock_device)
   subscribe_request:merge(DoorLock.attributes.OperatingMode:subscribe(mock_device))
@@ -62,26 +68,22 @@ local function test_init()
   subscribe_request:merge(DoorLock.events.LockOperation:subscribe(mock_device))
   subscribe_request:merge(DoorLock.events.DoorLockAlarm:subscribe(mock_device))
   subscribe_request:merge(DoorLock.events.LockUserChange:subscribe(mock_device))
-  -- add test device, handle initial subscribe
-  test.mock_device.add_test_device(mock_device)
   test.socket["matter"]:__expect_send({mock_device.id, subscribe_request})
-  -- actual onboarding flow
-  test.socket.device_lifecycle:__queue_receive({ mock_device.id, "added" })
-  test.socket.capability:__expect_send(
-    mock_device:generate_test_message("main", capabilities.lockAlarm.alarm.clear({state_change = true}))
-  )
+  -- device_init
   test.socket.device_lifecycle:__queue_receive({ mock_device.id, "init" })
-  test.socket["matter"]:__expect_send({mock_device.id, subscribe_request})
+  -- do_configure
   test.socket.device_lifecycle:__queue_receive({ mock_device.id, "doConfigure" })
   test.socket.capability:__expect_send(
     mock_device:generate_test_message("main", capabilities.lock.supportedLockCommands({"lock", "unlock"}, {visibility = {displayed = false}}))
   )
   mock_device:expect_metadata_update({ profile = "lock-user-pin-schedule" })
   mock_device:expect_metadata_update({ provisioning_state = "PROVISIONED" })
+  -- info_changed
+  test.socket.device_lifecycle:__queue_receive(mock_device:generate_info_changed({ profile = t_utils.get_profile_definition("lock-user-pin-schedule.yml")}))
+  test.socket["matter"]:__expect_send({mock_device.id, subscribe_request})
 end
 
 test.set_test_init_function(test_init)
-
 
 test.register_coroutine_test(
   "Handle driverSwitched event",
@@ -972,23 +974,6 @@ test.register_message_test(
       ),
     }
   },
-  {
-     min_api_version = 17
-  }
-)
-
-test.register_coroutine_test(
-  "Added lifecycle event lock nocodes nobattery",
-  function()
-    test.socket.device_lifecycle:__queue_receive({ mock_device.id, "added" })
-
-    test.socket.capability:__expect_send(
-      mock_device:generate_test_message(
-        "main",
-        capabilities.lockAlarm.alarm.clear({state_change = true})
-      )
-    )
-  end,
   {
      min_api_version = 17
   }
@@ -2208,6 +2193,187 @@ test.register_coroutine_test(
   }
 )
 
+local LOCAL_START_TIME = "2025-01-15T09:00:00"
+local LOCAL_END_TIME = "2025-01-15T17:00:00"
+
+test.register_coroutine_test(
+  "Handle Add Year Day Schedule command received from SmartThings.",
+  function()
+    test.socket.capability:__queue_receive(
+      {
+        mock_device.id,
+        {
+          capability = capabilities.lockSchedules.ID,
+          command = "setYearDaySchedule",
+          args = {1, 1, {localStartTime=LOCAL_START_TIME, localEndTime=LOCAL_END_TIME}}
+        },
+      }
+    )
+    test.socket.matter:__expect_send(
+      {
+        mock_device.id,
+        DoorLock.server.commands.SetYearDaySchedule(
+          mock_device, 1, -- endpoint
+          1, -- Schedule Index
+          1, -- User Index
+          lock_utils.iso8601_to_epoch(LOCAL_START_TIME), -- Start Time (epoch)
+          lock_utils.iso8601_to_epoch(LOCAL_END_TIME) -- End Time (epoch)
+        ),
+      }
+    )
+    test.wait_for_events()
+    test.socket.matter:__queue_receive(
+      {
+        mock_device.id,
+        DoorLock.server.commands.SetYearDaySchedule:build_test_command_response(
+          mock_device, 1
+        )
+      }
+    )
+    test.socket.capability:__expect_send(
+      mock_device:generate_test_message(
+        "main",
+        capabilities.lockSchedules.yearDaySchedules(
+          {{
+            userIndex=1,
+            schedules={{
+              scheduleIndex = 1,
+              localStartTime = LOCAL_START_TIME,
+              localEndTime = LOCAL_END_TIME
+            }},
+          }},
+          {visibility={displayed=false}}
+        )
+      )
+    )
+    test.socket.capability:__expect_send(
+      mock_device:generate_test_message(
+        "main",
+        capabilities.lockSchedules.commandResult(
+          {commandName="setYearDaySchedule", userIndex=1, scheduleIndex=1, statusCode="success"},
+          {state_change=true, visibility={displayed=false}}
+        )
+      )
+    )
+  end,
+  {
+     min_api_version = 17
+  }
+)
+
+test.register_coroutine_test(
+  "Add Year Day Schedule command received from SmartThings and send busy state",
+  function()
+    mock_device:set_field(lock_utils.BUSY_STATE, os.time(), {persist = true})
+    test.socket.capability:__queue_receive(
+      {
+        mock_device.id,
+        {
+          capability = capabilities.lockSchedules.ID,
+          command = "setYearDaySchedule",
+          args = {1, 1, {localStartTime=LOCAL_START_TIME, localEndTime=LOCAL_END_TIME}}
+        },
+      }
+    )
+    test.socket.capability:__expect_send(
+      mock_device:generate_test_message(
+        "main",
+        capabilities.lockSchedules.commandResult(
+          {commandName="setYearDaySchedule", statusCode="busy"},
+          {state_change=true, visibility={displayed=false}}
+        )
+      )
+    )
+  end,
+  {
+     min_api_version = 17
+  }
+)
+
+test.register_coroutine_test(
+  "Handle Clear Year Day Schedule command received from SmartThings.",
+  function()
+    test.socket.capability:__queue_receive(
+      {
+        mock_device.id,
+        {
+          capability = capabilities.lockSchedules.ID,
+          command = "clearYearDaySchedules",
+          args = {
+            1, -- user index
+            1, -- schedule index
+          }
+        },
+      }
+    )
+    test.socket.matter:__expect_send(
+      {
+        mock_device.id,
+        DoorLock.server.commands.SetYearDaySchedule(
+          mock_device, 1, -- endpoint
+          1, -- year_day_index
+          1, -- user_index
+          0, -- local_start_time
+          lock_utils.MAX_EPOCH_S -- local_end_time
+        ),
+      }
+    )
+    test.wait_for_events()
+    test.socket.matter:__queue_receive(
+      {
+        mock_device.id,
+        DoorLock.server.commands.SetYearDaySchedule:build_test_command_response(
+          mock_device, 1
+        )
+      }
+    )
+    test.socket.capability:__expect_send(
+      mock_device:generate_test_message(
+        "main",
+        capabilities.lockSchedules.commandResult(
+          {commandName="clearYearDaySchedules", userIndex=1, scheduleIndex=1, statusCode="success"},
+          {state_change=true, visibility={displayed=false}}
+        )
+      )
+    )
+  end,
+  {
+     min_api_version = 17
+  }
+)
+
+test.register_coroutine_test(
+  "Clear Year Day Schedule command received from SmartThings and send busy state",
+  function()
+    mock_device:set_field(lock_utils.BUSY_STATE, os.time(), {persist = true})
+    test.socket.capability:__queue_receive(
+      {
+        mock_device.id,
+        {
+          capability = capabilities.lockSchedules.ID,
+          command = "clearYearDaySchedules",
+          args = {
+            1, -- user index
+            1, -- schedule index
+          }
+        },
+      }
+    )
+    test.socket.capability:__expect_send(
+      mock_device:generate_test_message(
+        "main",
+        capabilities.lockSchedules.commandResult(
+          {commandName="clearYearDaySchedules", statusCode="busy"},
+          {state_change=true, visibility={displayed=false}}
+        )
+      )
+    )
+  end,
+  {
+     min_api_version = 17
+  }
+)
+
 test.register_coroutine_test(
   "Add Guest User and failure response ",
   function()
@@ -2446,6 +2612,5 @@ test.register_coroutine_test(
      min_api_version = 17
   }
 )
-
 
 test.run_registered_tests()
